+++ app/src/main/java/algorithm/master/currencyconverterpro/domain/model/history/RateModel.kt
@@ -6,5 +6,5 @@ import java.util.*
  * Created by promasterguru on 15/11/2022.
  */
 data class RateModel(
-    val base: String, val to: String, val amount: Double, val date: String, val endDate: String
+    val base: String, val to: String, val amount: Float, val date: String, val endDate: String
 )
